+++ src/main/java/com/campus/campus/domain/councilpost/application/service/StudentCouncilPostService.java
@@ -3,7 +3,6 @@ package com.campus.campus.domain.councilpost.application.service;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -15,8 +14,6 @@ import org.springframework.transaction.annotation.Transactional;
 import com.campus.campus.domain.council.application.exception.StudentCouncilNotFoundException;
 import com.campus.campus.domain.council.domain.entity.StudentCouncil;
 import com.campus.campus.domain.council.domain.repository.StudentCouncilRepository;
-import com.campus.campus.domain.councilpost.application.dto.response.GetLikedPostResponse;
-import com.campus.campus.domain.councilpost.application.dto.response.LikePostResponse;
 import com.campus.campus.domain.councilpost.application.dto.response.GetPostListForCouncilResponse;
 import com.campus.campus.domain.councilpost.application.dto.response.GetUpcomingEventListForCouncilResponse;
 import com.campus.campus.domain.councilpost.application.dto.response.NormalizedDateTime;
@@ -28,15 +25,12 @@ import com.campus.campus.domain.councilpost.application.exception.PostNotFoundEx
 import com.campus.campus.domain.councilpost.application.exception.PostOciImageDeleteFailedException;
 import com.campus.campus.domain.councilpost.application.exception.ThumbnailRequiredException;
 import com.campus.campus.domain.councilpost.application.mapper.StudentCouncilPostMapper;
-import com.campus.campus.domain.councilpost.domain.entity.LikePost;
 import com.campus.campus.domain.councilpost.domain.entity.PostCategory;
 import com.campus.campus.domain.councilpost.domain.entity.PostImage;
 import com.campus.campus.domain.councilpost.domain.entity.StudentCouncilPost;
 import com.campus.campus.domain.councilpost.domain.repository.LikePostRepository;
 import com.campus.campus.domain.councilpost.domain.repository.PostImageRepository;
 import com.campus.campus.domain.councilpost.domain.repository.StudentCouncilPostRepository;
-import com.campus.campus.domain.user.application.exception.UserNotFoundException;
-import com.campus.campus.domain.user.domain.entity.User;
 import com.campus.campus.domain.user.domain.repository.UserRepository;
 import com.campus.campus.global.oci.application.service.PresignedUrlService;
 
@@ -149,42 +143,6 @@ public class StudentCouncilPostService {
 			now, limit, pageable);
 
 		return posts.map(studentCouncilPostMapper::toGetUpcomingEventListForCouncilResponse);
-	}
-
-	@Transactional
-	public LikePostResponse toggleLikePost(Long userId, Long postId) {
-		User user = userRepository.findByIdAndDeletedAtIsNull(userId)
-			.orElseThrow(UserNotFoundException::new);
-
-		StudentCouncilPost post = postRepository.findById(postId)
-			.orElseThrow(PostNotFoundException::new);
-
-		Optional<LikePost> likePost = likePostRepository.findByUserIdAndPost_Id(userId, postId);
-		boolean isLike;
-		if (likePost.isPresent()) {
-			likePostRepository.delete(likePost.get());
-			isLike = false;
-		} else {
-			LikePost newLikePost = studentCouncilPostMapper.createLikePost(user, post);
-			likePostRepository.save(newLikePost);
-			isLike = true;
-		}
-
-		return studentCouncilPostMapper.toLikePostResponse(user, post, isLike);
-	}
-
-	@Transactional(readOnly = true)
-	public Page<GetLikedPostResponse> findLikedPosts(PostCategory category, int page, int size, Long userId) {
-		userRepository.findByIdAndDeletedAtIsNull(userId)
-			.orElseThrow(UserNotFoundException::new);
-
-		Pageable pageable = PageRequest.of(Math.max(page - 1, 0), size);
-
-		Page<LikePost> likedPosts = likePostRepository.findLikedPosts(userId, category, pageable);
-
-		return likedPosts.map(likePost ->
-			studentCouncilPostMapper.toGetLikedPostResponse(likePost.getPost())
-		);
 	}
 
 	@Transactional
