+++ src/main/java/com/campus/campus/domain/councilpost/presentation/StudentCouncilPostForUserController.java
@@ -6,12 +6,15 @@ import org.springframework.data.domain.Page;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.campus.campus.domain.council.domain.entity.CouncilType;
 import com.campus.campus.domain.councilpost.application.dto.response.GetActivePartnershipListForUserResponse;
+import com.campus.campus.domain.councilpost.application.dto.response.GetLikedPostResponse;
+import com.campus.campus.domain.councilpost.application.dto.response.LikePostResponse;
 import com.campus.campus.domain.councilpost.application.dto.response.PostListItemResponse;
 import com.campus.campus.domain.councilpost.application.dto.response.PostResponse;
 import com.campus.campus.domain.councilpost.application.service.StudentCouncilPostForUserService;
@@ -31,6 +34,29 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 @Slf4j
 public class StudentCouncilPostForUserController {
+
+	@PostMapping("/{postId}/like")
+	@PreAuthorize("hasRole('USER')")
+	@Operation(summary = "학생회 게시글 좋아요 토글")
+	public CommonResponse<LikePostResponse> togglePostLike(@PathVariable Long postId, @CurrentUserId Long userId) {
+		LikePostResponse response = postService.toggleLikePost(userId, postId);
+
+		return CommonResponse.success(StudentCouncilPostResponseCode.POST_LIKE_SUCCESS, response);
+	}
+
+	@GetMapping("/likes")
+	@PreAuthorize("hasRole('USER')")
+	@Operation(summary = "관심 학생회 게시글 목록 조회")
+	public CommonResponse<Page<GetLikedPostResponse>> getLikedPosts(
+		@RequestParam(required = false) PostCategory category,
+		@RequestParam(defaultValue = "1") int page,
+		@RequestParam(defaultValue = "3") int size,
+		@CurrentUserId Long userId
+	) {
+		Page<GetLikedPostResponse> responses = postService.findLikedPosts(category, page, size, userId);
+
+		return CommonResponse.success(StudentCouncilPostResponseCode.POST_LIST_READ_SUCCESS, responses);
+	}
 
 	private final StudentCouncilPostForUserService postService;
